+++ angular-workspace/projects/ni/nimble-angular/table-column/enum-text/public-api.ts
@@ -1,2 +1,0 @@
-export * from './nimble-table-column-enum-text.directive';
-export * from './nimble-table-column-enum-text.module';