+++ packages/nimble-components/src/table-column/icon/tests/table-column-icon.spec.ts
@@ -18,6 +18,7 @@ import { mappingSpinnerTag } from '../../../mapping/spinner';
 import { spinnerTag } from '../../../spinner';
 import { themeProviderTag } from '../../../theme-provider';
 import { TableColumnIconPageObject } from '../testing/table-column-icon.pageobject';
+import { mappingUserTag } from '../../../mapping/user';
 
 interface SimpleTableRecord extends TableRecord {
     field1?: MappingKey | null;
@@ -35,6 +36,11 @@ interface BasicSpinnerMapping {
     key?: MappingKey;
     text?: string;
     textHidden?: boolean;
+}
+
+interface BasicTextMapping {
+    key?: MappingKey;
+    text?: string;
 }
 
 class Model {
@@ -55,8 +61,9 @@ describe('TableColumnIcon', () => {
     // prettier-ignore
     async function setup(options: {
         keyType: MappingKeyType,
-        iconMappings: BasicIconMapping[],
-        spinnerMappings: BasicSpinnerMapping[]
+        iconMappings?: BasicIconMapping[],
+        spinnerMappings?: BasicSpinnerMapping[],
+        textMappings?: BasicTextMapping[]
     }): Promise<ModelFixture<Table<SimpleTableRecord>>> {
         const source = new Model();
         const result = await fixture<Table<SimpleTableRecord>>(html<Model>`
@@ -64,7 +71,7 @@ describe('TableColumnIcon', () => {
                 <${tableTag} ${ref('table')} style="width: 700px">
                     <${tableColumnIconTag} ${ref('col1')} field-name="field1" key-type="${options.keyType}">
                         Column 1
-                        ${repeat(() => options.iconMappings, html<BasicIconMapping>`
+                        ${repeat(() => options.iconMappings ?? [], html<BasicIconMapping>`
                             <${mappingIconTag}
                                 key="${x => x.key}"
                                 text="${x => x.text}"
@@ -72,12 +79,18 @@ describe('TableColumnIcon', () => {
                                 ?text-hidden="${x => x.textHidden}">
                             </${mappingIconTag}>
                         `)}
-                        ${repeat(() => options.spinnerMappings, html<BasicSpinnerMapping>`
+                        ${repeat(() => options.spinnerMappings ?? [], html<BasicSpinnerMapping>`
                         <${mappingSpinnerTag}
                             key="${x => x.key}"
                             text="${x => x.text}"
                             ?text-hidden="${x => x.textHidden}">
                         </${mappingSpinnerTag}>
+                        `)}
+                        ${repeat(() => options.textMappings ?? [], html<BasicTextMapping>`
+                        <${mappingTextTag}
+                            key="${x => x.key}"
+                            text="${x => x.text}"
+                        </${mappingTextTag}>
                         `)}
                     </${tableColumnIconTag}>
                 </${tableTag}>
@@ -116,8 +129,7 @@ describe('TableColumnIcon', () => {
                     keyType: value.name,
                     iconMappings: [
                         { key: value.key, text: 'alpha', icon: iconXmarkTag }
-                    ],
-                    spinnerMappings: []
+                    ]
                 }));
                 pageObject = new TablePageObject<SimpleTableRecord>(
                     model.table
@@ -137,7 +149,6 @@ describe('TableColumnIcon', () => {
             spec(`displays spinner mapped from ${name}`, async () => {
                 ({ connect, disconnect, model } = await setup({
                     keyType: value.name,
-                    iconMappings: [],
                     spinnerMappings: [{ key: value.key, text: 'alpha' }]
                 }));
                 pageObject = new TablePageObject<SimpleTableRecord>(
@@ -158,8 +169,7 @@ describe('TableColumnIcon', () => {
     it('displays blank when no matches', async () => {
         ({ connect, disconnect, model } = await setup({
             keyType: MappingKeyType.string,
-            iconMappings: [{ key: 'a', text: 'alpha', icon: iconXmarkTag }],
-            spinnerMappings: []
+            iconMappings: [{ key: 'a', text: 'alpha', icon: iconXmarkTag }]
         }));
         pageObject = new TablePageObject<SimpleTableRecord>(model.table);
         columnPageObject = new TableColumnIconPageObject(pageObject);
@@ -173,8 +183,7 @@ describe('TableColumnIcon', () => {
     it('displays blank when no icon specified for mapping', async () => {
         ({ connect, disconnect, model } = await setup({
             keyType: MappingKeyType.string,
-            iconMappings: [{ key: 'a', text: 'alpha', icon: undefined }],
-            spinnerMappings: []
+            iconMappings: [{ key: 'a', text: 'alpha', icon: undefined }]
         }));
         pageObject = new TablePageObject<SimpleTableRecord>(model.table);
         columnPageObject = new TableColumnIconPageObject(pageObject);
@@ -191,8 +200,7 @@ describe('TableColumnIcon', () => {
             iconMappings: [
                 { key: 'a', text: 'alpha', icon: iconXmarkTag },
                 { key: 'b', text: 'bravo', icon: iconCheckTag }
-            ],
-            spinnerMappings: []
+            ]
         }));
         pageObject = new TablePageObject<SimpleTableRecord>(model.table);
         columnPageObject = new TableColumnIconPageObject(pageObject);
@@ -211,8 +219,7 @@ describe('TableColumnIcon', () => {
     it('changing mapping icon updates display', async () => {
         ({ connect, disconnect, model } = await setup({
             keyType: MappingKeyType.string,
-            iconMappings: [{ key: 'a', text: 'alpha', icon: iconXmarkTag }],
-            spinnerMappings: []
+            iconMappings: [{ key: 'a', text: 'alpha', icon: iconXmarkTag }]
         }));
         pageObject = new TablePageObject<SimpleTableRecord>(model.table);
         columnPageObject = new TableColumnIconPageObject(pageObject);
@@ -232,8 +239,7 @@ describe('TableColumnIcon', () => {
     it('changing mapping severity updates display', async () => {
         ({ connect, disconnect, model } = await setup({
             keyType: MappingKeyType.string,
-            iconMappings: [{ key: 'a', text: 'alpha', icon: iconXmarkTag }],
-            spinnerMappings: []
+            iconMappings: [{ key: 'a', text: 'alpha', icon: iconXmarkTag }]
         }));
         pageObject = new TablePageObject<SimpleTableRecord>(model.table);
         columnPageObject = new TableColumnIconPageObject(pageObject);
@@ -253,8 +259,7 @@ describe('TableColumnIcon', () => {
     it('changing mapping key updates display', async () => {
         ({ connect, disconnect, model } = await setup({
             keyType: MappingKeyType.string,
-            iconMappings: [{ key: 'a', text: 'alpha', icon: iconXmarkTag }],
-            spinnerMappings: []
+            iconMappings: [{ key: 'a', text: 'alpha', icon: iconXmarkTag }]
         }));
         pageObject = new TablePageObject<SimpleTableRecord>(model.table);
         columnPageObject = new TableColumnIconPageObject(pageObject);
@@ -282,8 +287,7 @@ describe('TableColumnIcon', () => {
                             text: name,
                             icon: iconXmarkTag
                         }
-                    ],
-                    spinnerMappings: []
+                    ]
                 }));
                 pageObject = new TablePageObject<SimpleTableRecord>(
                     model.table
@@ -305,8 +309,7 @@ describe('TableColumnIcon', () => {
     it('sets group header text label and no icon when icon is undefined', async () => {
         ({ connect, disconnect, model } = await setup({
             keyType: MappingKeyType.string,
-            iconMappings: [{ key: 'b', text: 'bravo', icon: undefined }],
-            spinnerMappings: []
+            iconMappings: [{ key: 'b', text: 'bravo', icon: undefined }]
         }));
         pageObject = new TablePageObject<SimpleTableRecord>(model.table);
         columnPageObject = new TableColumnIconPageObject(pageObject);
@@ -323,8 +326,7 @@ describe('TableColumnIcon', () => {
     it('clears cell when mappings removed', async () => {
         ({ connect, disconnect, model } = await setup({
             keyType: MappingKeyType.string,
-            iconMappings: [{ key: 'a', text: 'alpha', icon: iconXmarkTag }],
-            spinnerMappings: []
+            iconMappings: [{ key: 'a', text: 'alpha', icon: iconXmarkTag }]
         }));
         pageObject = new TablePageObject<SimpleTableRecord>(model.table);
         columnPageObject = new TableColumnIconPageObject(pageObject);
@@ -343,8 +345,7 @@ describe('TableColumnIcon', () => {
     it('clears group header when mappings removed', async () => {
         ({ connect, disconnect, model } = await setup({
             keyType: MappingKeyType.string,
-            iconMappings: [{ key: 'a', text: 'alpha', icon: iconXmarkTag }],
-            spinnerMappings: []
+            iconMappings: [{ key: 'a', text: 'alpha', icon: iconXmarkTag }]
         }));
         pageObject = new TablePageObject<SimpleTableRecord>(model.table);
         columnPageObject = new TableColumnIconPageObject(pageObject);
@@ -364,9 +365,7 @@ describe('TableColumnIcon', () => {
     describe('validation', () => {
         it('is valid with no mappings', async () => {
             ({ connect, disconnect, model } = await setup({
-                keyType: MappingKeyType.number,
-                iconMappings: [],
-                spinnerMappings: []
+                keyType: MappingKeyType.number
             }));
             await connect();
             await waitForUpdatesAsync();
@@ -388,8 +387,7 @@ describe('TableColumnIcon', () => {
                     { key: '1.01', text: 'alpha', icon: iconXmarkTag },
                     { key: '-1.01', text: 'alpha', icon: iconXmarkTag },
                     { key: '-1e3', text: 'alpha', icon: iconXmarkTag }
-                ],
-                spinnerMappings: []
+                ]
             }));
             await connect();
             await waitForUpdatesAsync();
@@ -415,8 +413,7 @@ describe('TableColumnIcon', () => {
                                 text: 'alpha',
                                 icon: iconXmarkTag
                             }
-                        ],
-                        spinnerMappings: []
+                        ]
                     }));
                     await connect();
                     await waitForUpdatesAsync();
@@ -431,8 +428,7 @@ describe('TableColumnIcon', () => {
         it('is invalid with invalid numeric key values', async () => {
             ({ connect, disconnect, model } = await setup({
                 keyType: MappingKeyType.number,
-                iconMappings: [{ key: 'a', text: 'alpha', icon: iconXmarkTag }],
-                spinnerMappings: []
+                iconMappings: [{ key: 'a', text: 'alpha', icon: iconXmarkTag }]
             }));
             await connect();
             await waitForUpdatesAsync();
@@ -448,13 +444,13 @@ describe('TableColumnIcon', () => {
                 html`<${tableTag} style="width: 700px">
                         <${tableColumnIconTag} field-name="field1">
                             Column 1
-                            <${mappingTextTag} key="foo" text="foo"></${mappingTextTag}>
-                            <${mappingIconTag} key="bar" text="bar" icon="nimble-icon-xmark"></${mappingIconTag}>
+                            <${mappingUserTag} key="foo"></${mappingUserTag}>
+                            <${mappingUserTag} key="bar"></${mappingUserTag}>
                         </${tableColumnIconTag}>
                     </${tableTag}>`
             );
         }
-        it('is invalid with text mapping', async () => {
+        it('is invalid with unsupported mapping', async () => {
             let element: Table<SimpleTableRecord>;
             ({ element, connect, disconnect } = await setupInvalidMappings());
             await connect();
@@ -470,8 +466,7 @@ describe('TableColumnIcon', () => {
                 iconMappings: [
                     { key: 'a', text: 'alpha', icon: iconXmarkTag },
                     { key: 'a', text: 'alpha', icon: iconXmarkTag }
-                ],
-                spinnerMappings: []
+                ]
             }));
             await connect();
             await waitForUpdatesAsync();
@@ -485,8 +480,7 @@ describe('TableColumnIcon', () => {
                 iconMappings: [
                     { key: '0', text: 'alpha', icon: iconXmarkTag },
                     { key: '0.0', text: 'alpha', icon: iconXmarkTag }
-                ],
-                spinnerMappings: []
+                ]
             }));
             await connect();
             await waitForUpdatesAsync();
@@ -494,11 +488,10 @@ describe('TableColumnIcon', () => {
             expect(model.col1.validity.duplicateMappingKey).toBeTrue();
         });
 
-        it('is invalid with missing key value', async () => {
+        it('is invalid with missing icon key value', async () => {
             ({ connect, disconnect, model } = await setup({
                 keyType: MappingKeyType.string,
-                iconMappings: [{ text: 'alpha', icon: iconXmarkTag }],
-                spinnerMappings: []
+                iconMappings: [{ text: 'alpha', icon: iconXmarkTag }]
             }));
             await connect();
             await waitForUpdatesAsync();
@@ -509,8 +502,7 @@ describe('TableColumnIcon', () => {
         it('is invalid with missing icon text value', async () => {
             ({ connect, disconnect, model } = await setup({
                 keyType: MappingKeyType.string,
-                iconMappings: [{ key: 'a', icon: iconXmarkTag }],
-                spinnerMappings: []
+                iconMappings: [{ key: 'a', icon: iconXmarkTag }]
             }));
             await connect();
             await waitForUpdatesAsync();
@@ -521,8 +513,7 @@ describe('TableColumnIcon', () => {
         it('is invalid with non-icon icon value', async () => {
             ({ connect, disconnect, model } = await setup({
                 keyType: MappingKeyType.string,
-                iconMappings: [{ key: 'a', text: 'alpha', icon: 'div' }],
-                spinnerMappings: []
+                iconMappings: [{ key: 'a', text: 'alpha', icon: 'div' }]
             }));
             await connect();
             await waitForUpdatesAsync();
@@ -533,8 +524,7 @@ describe('TableColumnIcon', () => {
         it('is invalid with completely made up icon value', async () => {
             ({ connect, disconnect, model } = await setup({
                 keyType: MappingKeyType.string,
-                iconMappings: [{ key: 'a', text: 'alpha', icon: 'foo' }],
-                spinnerMappings: []
+                iconMappings: [{ key: 'a', text: 'alpha', icon: 'foo' }]
             }));
             await connect();
             await waitForUpdatesAsync();
@@ -545,13 +535,45 @@ describe('TableColumnIcon', () => {
         it('is invalid with missing spinner text value', async () => {
             ({ connect, disconnect, model } = await setup({
                 keyType: MappingKeyType.string,
-                iconMappings: [],
                 spinnerMappings: [{ key: 'a' }]
             }));
             await connect();
             await waitForUpdatesAsync();
             expect(model.col1.checkValidity()).toBeFalse();
             expect(model.col1.validity.missingTextValue).toBeTrue();
+        });
+
+        it('is invalid with missing spinner key value', async () => {
+            ({ connect, disconnect, model } = await setup({
+                keyType: MappingKeyType.string,
+                spinnerMappings: [{ text: 'alpha' }]
+            }));
+            await connect();
+            await waitForUpdatesAsync();
+            expect(model.col1.checkValidity()).toBeFalse();
+            expect(model.col1.validity.missingKeyValue).toBeTrue();
+        });
+
+        it('is invalid with missing spinner text value', async () => {
+            ({ connect, disconnect, model } = await setup({
+                keyType: MappingKeyType.string,
+                textMappings: [{ key: 'a' }]
+            }));
+            await connect();
+            await waitForUpdatesAsync();
+            expect(model.col1.checkValidity()).toBeFalse();
+            expect(model.col1.validity.missingTextValue).toBeTrue();
+        });
+
+        it('is invalid with missing text key value', async () => {
+            ({ connect, disconnect, model } = await setup({
+                keyType: MappingKeyType.string,
+                textMappings: [{ text: 'alpha' }]
+            }));
+            await connect();
+            await waitForUpdatesAsync();
+            expect(model.col1.checkValidity()).toBeFalse();
+            expect(model.col1.validity.missingKeyValue).toBeTrue();
         });
     });
 
@@ -583,7 +605,6 @@ describe('TableColumnIcon', () => {
             spec(`group row renders expected value when ${name}`, async () => {
                 ({ connect, disconnect, model } = await setup({
                     keyType: MappingKeyType.string,
-                    iconMappings: [],
                     spinnerMappings: [{ key: 'a', text: 'a' }]
                 }));
                 pageObject = new TablePageObject<SimpleTableRecord>(
@@ -717,6 +738,10 @@ describe('TableColumnIcon', () => {
             {
                 name: 'icon mapping',
                 type: 'icon'
+            },
+            {
+                name: 'text mapping',
+                type: 'text'
             }
         ] as const;
 
@@ -746,6 +771,10 @@ describe('TableColumnIcon', () => {
                         spinnerMappings: [
                             { key: 'spinner-long', text: longText },
                             { key: 'spinner-short', text: shortText }
+                        ],
+                        textMappings: [
+                            { key: 'text-long', text: longText },
+                            { key: 'text-short', text: shortText }
                         ]
                     }));
                     pageObject = new TablePageObject<SimpleTableRecord>(